Hotfix: Change bathrooms type to float in PropertyResponse to fix Pydantic validation error

+++ backend/routers/properties.py
@@ -12,7 +12,7 @@ class PropertyResponse(BaseModel):
     price: float
     sqft: float
     bedrooms: int
-    bathrooms: int
+    bathrooms: float
     latitude: float
     longitude: float
